# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PreviewDetail": {
        "properties": {
          "comparison": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comparison"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "input_hash": {
            "title": "Input Hash",
            "type": "string"
          },
          "inputs": {
            "additionalProperties": true,
            "title": "Inputs",
            "type": "object"
          },
          "parent_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "preview_id": {
            "format": "uuid",
            "title": "Preview Id",
            "type": "string"
          },
          "result": {
            "additionalProperties": true,
            "title": "Result",
            "type": "object"
          },
          "result_hash": {
            "title": "Result Hash",
            "type": "string"
          }
        },
        "required": [
          "preview_id",
          "parent_id",
          "created_at",
          "input_hash",
          "result_hash",
          "inputs",
          "result",
          "comparison"
        ],
        "title": "PreviewDetail",
        "type": "object"
      },
      "PreviewList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PreviewSummary"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "PreviewList",
        "type": "object"
      },
      "PreviewRequest": {
        "additionalProperties": false,
        "properties": {
          "inputs": {
            "additionalProperties": true,
            "title": "Inputs",
            "type": "object"
          },
          "parent_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          }
        },
        "required": [
          "inputs"
        ],
        "title": "PreviewRequest",
        "type": "object"
      },
      "PreviewSummary": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "input_hash": {
            "title": "Input Hash",
            "type": "string"
          },
          "parent_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "preview_id": {
            "format": "uuid",
            "title": "Preview Id",
            "type": "string"
          },
          "result_hash": {
            "title": "Result Hash",
            "type": "string"
          }
        },
        "required": [
          "preview_id",
          "parent_id",
          "created_at",
          "input_hash",
          "result_hash"
        ],
        "title": "PreviewSummary",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Equipment Isolation Agent API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/planning-previews": {
      "get": {
        "operationId": "list_previews_planning_previews_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewList"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Previews",
        "tags": [
          "Synthetic planning previews"
        ]
      },
      "post": {
        "operationId": "create_preview_planning_previews_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Preview",
        "tags": [
          "Synthetic planning previews"
        ]
      }
    },
...
    "/planning-previews/{preview_id}": {
      "get": {
        "operationId": "read_preview_planning_previews__preview_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "preview_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Preview Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Preview",
        "tags": [
          "Synthetic planning previews"
        ]
      }
    },
...
  }
}